Move get_models
Move get_models from tests.paired_fields_utils to jobserver.model_utils.
This function is used outside testing paired fields and, whilst it isn't
used outside testing, adding a tests.model_utils as well as a
jobserver.model_utils seemed excessive.
